Dictionary question
(1 pts) Create a dictionary. Its keys consist of students' names and the values consist of the list of students' grades.
(2 pts) Take both the student names and the student grades from the user.
(3 pts) Continue taking students' names until a user enters -1.
(3 pts) Each time, check whether a user enter the same name before. (In this class, everyone has a unique name.)
(2 pts) Take exactly three grades for each student.
(3pts) Each grade should be a positive number including zero, please check each grade whether it meets this condition.
(3pts) Calculate the average grade for each student and print it on the screen.
(3pts) Check the dictionary below, which shows the grading scale, and determine the letter grade of each student based on it. Add the letter grades of the students at the end of the student's grade lists and print the final version of the dictionary.
grading_scale = {'A': '100-90', 'B': '89-70', 'C': '69-60', 'D': '59-50', 'F': '49-0'}